requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.135.2
filetype==1.2.0
google-auth==2.49.1
google-genai==1.68.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.13
langchain-core==1.2.22
langchain-google-genai==4.2.1
langchain-openai==1.1.12
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
openai==2.30.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pdf2image==1.17.0
pillow==12.1.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.2.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.28
requests==2.33.0
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
